Remove mysqlclient as a development dependency.
The mysqlclient development dependency was added a long time ago in
order to ensure that our custom MySQL backend (which fixes bugs in
Django 1.6-1.10's implementation) was being tested. However, that
dependency has been difficult to install on some systems, and the
bottleneck just isn't worth it anymore.

This is a trivial change that removes it. We'll continue to require it
for CI later.