Fix rendering extension static media bundles when using only source files.
When rendering an extension's static media files in debug mode, the
resulting HTML is returned as unsafe text. This can lead Django to
escape the contents, breaking pages and showing HTML source in the
pages.

This fixes this by explicitly marking the HTML as safe. The bulk of the
change, though, is to the unit tests. We now have two versions of most
of the tests: One with Pipeline enabled, one with Pipeline disabled.
This gives us full test coverage of this logic so that neither mode can
regress without tests failing in the future.

Testing Done:
Unit tests pass on all supported versions of Django and Python.

Reviewed at https://reviews.reviewboard.org/r/10822/