Fix a legacy call to cycle in the log template.
We have a call to `{% cycle %}` in the log template that uses legacy
syntax, which isn't supported on Django 1.11. This is a simple update
that switches it to use the modern syntax, compatible on both 1.6 and
1.11.

Testing Done:
Tested the log viewer. It rendered without breaking on Django 1.11.

Reviewed at https://reviews.reviewboard.org/r/10839/