Focus modal box upon initialization
This change makes the modal box focusable by adding the attribute
`tabindex="0"` to the element and calling focus on it when it is
initialized. This allows views that have `modalBox` as a parent to
detect key events within it, such as `RB.DialogView`.

Testing Done:
Manual verification that the modal box is focused when it is created in
the new review request dialog and the file upload dialog. This is
indicated by the focus ring around the dialogs in the attached
screenshots.

Reviewed at https://reviews.reviewboard.org/r/10992/