Use the new localization syntax for all ES6 JavaScript.
This makes use of the new babel-plugin-django-gettext template literal
support in our ES6 JavaScript files, giving us much more readable code,
especially when long strings are concerned.

We can now use the new `_` prefix and break strings into multiple lines,
and the plugin will take care to clean up the whitespace and turn them
back into single-line `gettext()` calls. We can also reference variables
in template literals, and the plugin will turn them into `interpolate()`
calls (though we don't make use of that currently int his codebase).

There are still files that aren't converted to ES6 in this codebase, and
there are likely a lot of unlocalized strings, but this change doesn't
tackle either. It's really the changes that were being tested during the
development of the plugin.

Testing Done:
Built a package and compared the generated code to the source code.
Saw that each and every new-style localization was converted to an
old-fashioned `gettext()`, with the expected strings.

Reviewed at https://reviews.reviewboard.org/r/11001/