Add an improved string representation for IntegrationConfig.
The `IntegrationConfig` string representation wasn't very nice. It
didn't list the configuration name, and it used the integration ID
directly. Not exactly friendly.

This improves it to return a string in the form of "Name (Integration)",
which is a bit more useful and friendly all-around.

Testing Done:
Checked this in the admin UI in places where it showed up.

Tested with configurations that mapped to an available integration
and ones that mapped to an unavailable integration (making sure that the
integration ID fallback worked).

Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/11015/