Update unit tests to not use fat arrow functions for Jasmine methods.
When we first moved the codebase to ES6, we were using fat arrow
functions in calls to `suite()`, `describe()`, `it()`, and
`beforeEach()`, which appeared to make sense at the time. However,
Jasmine actually intentionally provides an explicit `this` value used
for test context. We don't really use it in practice, but we don't want
to override it, and we're not supposed to be passing in fat arrow
functions to any of these calls.

This change simply replaces all `=> { ... }` in Jasmine calls with
`function() { ... }`.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/11010/