Update some Djblets components to use the new spinner.
This updates both the config forms `ListItemView` and the integrations
page to utilize the new `djblets-o-spinner` component, and to make use
of ARIA attributes to hint to screen readers that some areas of the page
are loading.

This just leaves the extensions page left using the old spinner, but
that's intentional, as that's being completely rewritten separately.

Testing Done:
Verified that the attributes were set correctly whenever the spinner
was shown for list items, and that the integrations page had the correct
attributes and spinner as well.

Reviewed at https://reviews.reviewboard.org/r/11035/