Add simple CSS components for actions above a list.
This helps to standardize the display of actions above a list through a
new `djblets-c-config-forms-list-actions` component. This is
intentionally very simple, just providing a basic set of classes to use
that can be styled or queried in a consistent way by a consuming
application. These are likely going to be mixed in with other classes
(such as `djblets-l-config-forms-container` for the `list-actions` and
something like `btn` for the `djblets-c-config-forms-list-action`).

The integrations list (currently the only part of Djblets using this
concept) has been updated to incorporate these classes.

Testing Done:
Made sure the styles didn't impact anything on the Integrations page.

Tested this with an upcoming change to the Extensions page.

Reviewed at https://reviews.reviewboard.org/r/11043/