Fix clicking on an extension's "Configure" button.
The Configure action for extensions wasn't working, raising an error in
the JS console about attempting to call `undefined`. The way list item
actions work, if there's a handler defined it will attempt to call it,
but otherwise it will just act as a link. The inclusion of the handler
name for `configure` inside the `actionHandlers` object meant it was
trying the former, but we never had defined the method it referenced.
Removing this entry makes it a link again.

Testing Done:
Clicked "Configure" on an extension and saw it open the extension's
configuration UI.

Reviewed at https://reviews.reviewboard.org/r/11117/