Log the "Reloading logging settings" as a debug log.
When the logging settings reload, we log a statement (as `info`) saying
the settings are reloading. This can be seen many times in a log, since
it'll often happen across multiple web server threads within seconds.

That's not entirely useful for most people to see, and tends to confuse
some, so we're now logging as a `debug` statement instead of an `info`
statement.

Testing Done:
Toggled logging settings and saw that this log statement was logged
as debug.

Reviewed at https://reviews.reviewboard.org/r/11116/