Fix clicking on an extension's "Database" link.
Like the recent change to fix the "Configure" link on the extensions
list, this change fixes clicking on the "Database" link. The issue here
was a bad entry in the handlers object, which overrides the default
link-like behavior.

Testing Done:
Opened the extensions page and clicked on the "Database" entry for an
extension which has database models.

Reviewed at https://reviews.reviewboard.org/r/11133/