Fix color of dashboard customization menu items.
The dashboard customization menu uses a bunch of `a` tags for each menu
item. In this case, a generic rule for `a:link` was taking higher
priority than the `.datagrid-menu a` rule, overriding the black color
with a blue. This change updates the djblets datagrid rule to be just as
specific.

Testing Done:
Looked at the dashboard customization menu, saw that the items therein
were black instead of blue.

Reviewed at https://reviews.reviewboard.org/r/11220/