Fix integration popup not disappearing on button click
In RB 4.0.x, the popup that comes up after selecting an integration's
`Delete` button fails to disappear after deleting the integration and
after selecting the cancel button. The issue was that in RB 3.0.x, we
are using `<input type="button">` to construct the buttons and from RB
4.0.x and onwards we are using the `<button>` tags instead. I've added a
condition so that the popup can be destroyed even if the `<button>` tags
are used.

Testing Done:
Manual testing done with the following repo pairs:
  - Tested Djblets release-2.0.x with RB release 4.0.x
  - Tested Djblets master with RB master

Reviewed at https://reviews.reviewboard.org/r/11276/