Enable universal wheels, and remove older packaging aliases.
Now that Djblets supports Python 3, it's time to turn on universal
wheels. This change enables this option, giving us a wheel that can be
used on either Python 2 or 3, removes older aliases (everything but
`release`), and adds a default `.dev` tag for builds. This is a copy of
what we have elsewhere.