Update the actual config forms page code to use modern config forms CSS.
Djblets 2.0 got a major update to config forms, moving to our modern CSS
component model and fixing up a bunch of styling. Everything in Djblets
was updated to use this except, ironically, the module that introduced
config forms in the first place.

The actual `configforms` Python module and related JavaScript code is
responsible for pages like the My Account page in Review Board, or the
admin UI in RBCommons. This module was still using legacy naming, which
meant we weren't able to get some of the benefits of newer styling, and
were building UIs that needed workarounds for layout.

This change moves this last remaining part of Djblets to the modern
components.

Testing Done:
Unit tests pass.

Tested each page in My Account

Reviewed at https://reviews.reviewboard.org/r/11331/